refactor: 팔로우 도메인 패키지 이동
