+++ src/main/java/com/example/sns/member/domain/FollowRepository.java
@@ -1,6 +1,5 @@
-package com.example.sns.social.domain;
+package com.example.sns.member.domain;
 
-import com.example.sns.member.domain.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
