+++ src/main/java/com/example/sns/social/application/SocialService.java
@@ -5,8 +5,8 @@ import com.example.sns.member.domain.MemberRepository;
 import com.example.sns.member.exception.MemberNotFoundException;
 import com.example.sns.social.application.dto.FollowRequest;
 import com.example.sns.social.application.dto.UnfollowRequest;
-import com.example.sns.social.domain.Follow;
-import com.example.sns.social.domain.FollowRepository;
+import com.example.sns.member.domain.Follow;
+import com.example.sns.member.domain.FollowRepository;
 import com.example.sns.social.exception.AlreadyFollowException;
 import com.example.sns.social.exception.NotFollowingMemberException;
 import lombok.RequiredArgsConstructor;
